fix for lcore > 255

+++ app/pktgen-main.c
@@ -537,7 +537,7 @@ main(int argc, char **argv)
 void
 pktgen_stop_running(void)
 {
-	uint8_t lid;
+	uint16_t lid;
 
 	for (lid = 0; lid < RTE_MAX_LCORE; lid++)
 		pg_stop_lcore(pktgen.l2p, lid);
